Here is the modified InputServer.cpp file which includes the casting change for the DispatchEvent() function.  This will allow the input_server to work with the current proto6 code in the cvs tree.  It may even work with the existing proto5.
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@297 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/servers/input/InputServer.cpp
@@ -938,8 +938,8 @@ int InputServer::DispatchEvent(BMessage *message)
     			int64 time=(int64)real_time_clock();
     			appsvrlink->SetOpCode(B_MOUSE_MOVED);
     			appsvrlink->Attach(&time,sizeof(int64));
-    			appsvrlink->Attach(&xValue,sizeof(float));
-    			appsvrlink->Attach(&yValue,sizeof(float));
+    			appsvrlink->Attach((float)&xValue,sizeof(float));
+    			appsvrlink->Attach((float)&yValue,sizeof(float));
     			message->FindInt32("buttons",buttons);
     			appsvrlink->Attach(&buttons,sizeof(int32));
     			appsvrlink->Flush();
